+++ src/server/api/mod.rs
@@ -1,6 +1,6 @@
 //! This module contains the API endpoints for the server.
+pub mod archive;
 pub mod routes;
 pub mod serve;
 pub mod state;
-pub mod stelae;
 pub mod versions;